repository/src/main/java/org/lfenergy/compas/scl/data/model/Item.java
// SPDX-FileCopyrightText: 2021 Alliander N.V.
//
// SPDX-License-Identifier: Apache-2.0
package org.lfenergy.compas.scl.data.model;

import javax.xml.bind.annotation.XmlAccessType;
import javax.xml.bind.annotation.XmlAccessorType;
import javax.xml.bind.annotation.XmlElement;
import javax.xml.bind.annotation.XmlRootElement;

@XmlRootElement(name = "Item")
@XmlAccessorType(XmlAccessType.FIELD)
public class Item {
    @XmlElement(
            name = "Id",
            required = true
    )
    private String id;
    @XmlElement(
            name = "Filename",
            required = true
    )
    private String filename;
    @XmlElement(
            name = "Version",
            required = true
    )
    private String version;

    public Item() {
    }

    public Item(String id, String filename, String version) {
        this.id = id;
        this.filename = filename;
        this.version = version;
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getFilename() {
        return filename;
    }

    public void setFilename(String filename) {
        this.filename = filename;
    }

    public String getVersion() {
        return version;
    }

    public void setVersion(String version) {
        this.version = version;
    }
}
